Fixed incorrect initialization of handler::pushed_cond.
The member field pushed_cond in class handler was only initialized to NULL in sql_select.cc
'if (tmp || !cond || tab->type == JT_REF)'. This caused some EXPLAINs to incorrectly report a 
pushed condition in cases where there actually was no conditions at all.

Also corrected explain results which was affected by this bug and removed a redundant MTR testcase
which was a direct copy of the preceeding testcase.