Bug#52092 Bad result in 6.0-codebase after setting correct block_size ...
Problem: LooseScan strategy fails because the partition engine does not
deliver the rows in sorted index order. This happens because when a 
a temporary table is used, the requirement for ordered index access
is turned off.  (Since sorting will have to be done later, anyway).  
However, LooseScan strategy needs ordered index access in order to work
correctly. 

Solution: Keep requirement for sorted access when LooseScan is to be used.