Fix possible waits when dropping an index
Fix possible waits when dropping an index similarly to the way they
were fixed in DROP TABLE - set trx->allowed_to_wait to FALSE during
que_eval_sql() so it does not wait on possible user locks but returns
DB_LOCK_WAIT_TIMEOUT immediately.

Also push the error message to the client in addition to printing it
to the server error log and change DROP TABLE not to fail if stats
cannot be deleted.