Bug#18144: Cost with FORCE/USE index seems incorrect in some cases.
Most of the times FORCE INDEX is ignored when the plan would be more expensive
than table scan. In GROUP BY cases, however, it is not, but the cost-based
optimizer used the same cost for performing a table scan by using a
non-covering index.
  
Fixed by setting the cost to the same as reading a single range the size of
the table.