Bug#26106: Wrong plan may be chosen when there are several possible range and ref accesses
For ref(const) access methods there are two kinds of statistics when a prefix
of a key is matched: range access estimate and table statistics, which are not
necessarily consistent with each other. This may make the optimizer choose a
key which matches fewer key parts than some other key, which is obviously not
correct.
  
Fixed by giving preferrence to the keys that match the most parts, and
choosing the best one among these.