Bug #48084 slave breaks if 'drop database' fails on master and mismatched tables on slave
'DROP TABLE <deleted tables>' was binlogged when 'DROP DATABASE' failed 
and at least one table was deleted from the database. The log event would
lead slave SQL thread stop if some of the tables did not exist on slave.
            
After this patch, It is always binlogged with 'IF EXISTS' option.