Followup for Bug #54360 Deadlock DROP/ALTER/CREATE DATABASE                         with open HANDLER
Fixes a problem with schema.test visible using embedded server.
The HANDLER was not closed which caused the test to hang.

The problem was not visible if the test was run on a normal server
as the the handler there was implicitly closed by DATABASE DDL
statements doing Events::drop_schema_events().