wl#2540 replication events checksum
testing with mysqlbinlog --remote revealed an issue with parsing checksummed fake Rotate
by the client.

It is safe to make the fake Rotate be sent to mysqlbinlog w/o checksum as on
mysqlbinlog side it does not play any significant role. In particular,
the real binary logged events checksum verification is not affected.