Refactored representation parent relationship in SPJ API 'Impl' classes.
Original SPJ API reflects the possibility for a NdbQueryOperation having multiple parents.
The possibilty for #parents > 1 has never been implemented, except that we used a Vector<> 
to hold the parent referenced for each QueryOperation - Several places there was an assert(#parents <= 1).

It is not liklely that we will support multiple parents in the foreseen future, and we has therefore
refactored parent refs to be a single NdbQueryOperation*.

The external SPJ API has been keep unchanged.