Fix a failing assert when running funcs_1.innodb_trig_03 test.
The failure was introduced by a precursor patch for the
fix for Bug#52044.

When opening tables for GRANT statement
to check that subject columns exist,
mysql_table_grant() would try to lock the
tables, and thus start a transaction.
This was unnecessary and lead to an assert.


sql/sql_acl.cc:
  Use open_normal_and_derived_tables() rather than
  open_and_lock_tables() to avoid an assert
  that no transaction is started in GRANT statement.