BUG#44209: MASTER_CONNECT_RETRY and --master-retry-count disconnected from each other
Post-push fix.

CHANGE MASTER TO master_retry_count=0 was not working properly, because
before assiging the value got from the lex, we would check if it was
different than zero. This is wrong.

To fix this we, deploy a flag that states when the value was indeed set
by a CHANGE MASTER TO command - which could even be 0 - and only then we
assign the new value.