Bug#51457: Firstmatch semijoin strategy gives wrong results for            certain query plans
When the optimizer selects a semijoin FirstMatch plan that has
non-correlated tables interspersed with inner tables, the "jumps"
are placed so that execution might become incorrect.
Example: A query with an outer correlated table ot, a non-correlated
table nt and two inner tables it1 and it2 are used in a query and the
optimizer chooses the FirstMatch join order ot - it1 - nt - it2.
The optimizer assigns a "jump" from table it2 to ot. This strategy will
omit duplicates resulting from table nt.

The fix assures that the jump from the last table in a consecutive
range of inner table goes to the most immediately preceeding outer
table (regardless of whether the outer table is correlated or not).
In this case, there are two consecutive ranges of inner tables:
{it1} and {it2}. In the first range, the last inner table is it1, the
jump should be to table ot. In the second range, the jump from it2
should be to nt.

It is currently possible to generate such query plans, hence the
bug fix includes some test cases. However, such plans are probably
generated due to a problem with the join optimizer cost model.
When this problem is fixed, it may be impossible to generate
such FirstMatch plans, unless one can assign different costs
to different tables, or another measure is used to force a specific
join order.

mysql-test/r/subselect_innodb.result
  Test results for bug#51457.

mysql-test/t/subselect_innodb.test
  Test case for bug#51457. Notice the above comment for test validity.
  The test goes into subselect_innodb because it is required to use
  tables with one row that are not hit by const table optimization.

sql/sql_select.cc
  In setup_semijoin_dups_elimination(), in the case for FIRST_MATCH,
  loop over the set of inner and non-correlated outer tables and assign
  jump target to the last table in each consecutive range of inner
  tables. The jump target should be the most immediately preceeding
  outer table.