Bug#32391 Character sets: crash with --character-set-server
Problem:
mysqld crashed on initialization of the built-in stopwords
when started with --character-set-server=utf16.
ft_init_stopwords() mistakenly compared the built-in 
stopwords using "utf16" as character set, which lead
to exit on DBUG_ASSERT((slen % 2) == 0) in
my_strnncollsp_utf16() when comparing a word with an even length
(for example word="about", len=5).

Fix:
Using latin1 when initializing the built-in stopwords   
for the "tricky" character sets.