This commit fixes two errors related to iterating over results from scan-scan queries  with multiple result batches. Both occur if the query tree has leaf lookup operations.
The first error was that two asserts that should only apply to scan operations
could also be triggered by lookups.

The second error meant that result rows would be lost because NdbResultStream
objects were erroneously reset even if descendant scans had more reults related
to the current batch (of the stream that was wrongky reset).

Also, a compiler warning is removed.