Bug #56241 Valgrind/CDB warnings for many tests in trunk-bugfixing
The reason for the warnings was that the performance schema
instrumentation function release_table_share() could be called
on a TABLE_SHARE that had already been free'd.

The bug was introduced during merging of the patch for Bug#52044
from mysql-5.5-bugfixing (which does not include the call to
PSI release_table_share()) to mysql-trunk-bugfixing.

This patch fixes the problem by moving the release_table_share()
call to TABLE_SHARE::destroy() before the TABLE_SHARE is free'd.

No test case added as this problem is detected by existing 
valgrind tests.