Fixed bug in SPJ block related to pruned scanIndex:
Only some of the 'fragCount' fragments will be active in a pruned indexscan. However,
we required that all fragCount fragments is 'done' in order to detect the scan as
such to be complete.

This fix will cause 'data.m_frags_complete' initially to also include #fragments which
never did receive a REQuest.