BUG#52344 - Subquery materialization:             Assertion if subquery in on-clause of outer join
Problem: If tables of an outer join are constant tables,
the associated on-clause will be evaluated in the optimization
phase.  If the on-clause contains a query that is to be
executed with subquery materialization, this will not work
since the infrastructure for such execution is not yet set up.

Solution: Do not evaluate on-clause in optimization phase if
is_expensive() returns true for this clause.  This is how the
problem is currently avoided for where-clauses.  This works
because, Item_in_subselect::is_expensive_processor returns true
if query is to be executed with subquery materialization.