BUG#53562: EXPLAIN statement should hint when index is not used            due to type conversion
An index may not be used for ref or range access if
 a) a type conversion is needed to compare an indexed 
    field to a value, or
 b) there is a collation mismatch between an indexed field 
    and the value compared to

Before, the index would not be used, but no hint 
was given that this was the case. With this patch, 
EXPLAIN EXTENDED will issue a warning for the cases above. 

Note that even if an index cannot be used for ref or range
access, the index may still be scanned.