The test cases rpl_*_crash_safe explicitly crashes the server by calling abort() that calls handle_segfault() and then exit(1).
While processing exit(1), associated handlers/callbacks are executed and destroy the
ACL_internal_schema_access and ACL_internal_table_access. Meanwhile, another thread
is able to start processing a statement and fails because the objects 
ACL_internal_schema_access and ACL_internal_table_access do not exist.

So the assertion in "cxa_pure_virtual()" fails because both destructors are called and
invalid objects are accessed.

This is not problem as there was a crash in the server and for that reason we simply
ignore the warning message.